+++ docs/src/conf.py
@@ -88,4 +88,5 @@ linkcheck_ignore = [
     # 404 Client Error: Not Found
     # See https://community.sonarsource.com/t/https-rules-sonarsource-com-down/177294?u=fniessink:
     "https://rules.sonarsource.com/.*",
+    "https://www.ictu.nl/en/about-us",  # False negative: 415 Client Error: Unsupported Media Type
 ]